pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "xplain"
version = "0.1.0"
description = "AI-powered CLI tool to explain code, commands, errors, and git diffs — powered by GitHub Models API"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "xplain Contributors" }
]
keywords = ["cli", "ai", "copilot", "developer-tools", "explain", "github-models"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
xplain = "src.cli:app"

[project.urls]
Homepage = "https://github.com/xplain-cli/xplain"
Repository = "https://github.com/xplain-cli/xplain"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
